change padding from pixels to percentage

+++ src/config.py
@@ -21,8 +21,8 @@ class Config:
     llm_extra_parameters: dict = field(default_factory=dict)
     font_min_size: int = 1
     font_max_size: int = 999
-    text_padding_h: int = 0
-    text_padding_v: int = 0
+    text_padding_h: float = 0.0
+    text_padding_v: float = 0.0
     line_spacing: float = 1.5
